feat: 유효성 검사 코드를 findBy~ 대신 exist~ 메서드를 사용하도록 수정
#273

+++ backend/emm-sale/src/main/java/com/emmsale/event/application/EventService.java
@@ -248,8 +248,9 @@ public class EventService {
   }
 
   public void deleteEvent(final Long eventId) {
-    eventRepository.findById(eventId)
-        .orElseThrow(() -> new EventException(NOT_FOUND_EVENT));
+    if (!eventRepository.existsById(eventId)) {
+      throw new EventException(NOT_FOUND_EVENT);
+    }
 
     eventRepository.deleteById(eventId);
   }
